refactor: organize
