+++ frontend/src/views/Dashboard/components/BCeIDBalance.jsx
